+++ console/src/feature/range/list/ContextMenu.tsx
@@ -8,16 +8,7 @@
 // included in the file licenses/APL.txt.
 
 import { ranger } from "@synnaxlabs/client";
-import {
-  Access,
-  Divider,
-  Icon,
-  List,
-  Menu,
-  Panel as PPanel,
-  Ranger,
-  Status,
-} from "@synnaxlabs/pluto";
+import { Access, Icon, List, Menu, Ranger, Status } from "@synnaxlabs/pluto";
 
 import { CreateChildRangeIcon } from "@/feature/range/ContextMenu";
 import { Cluster } from "@/platform/cluster";
@@ -49,7 +40,6 @@ export const ContextMenu = ({ keys }: Menu.ContextMenuMenuProps) => {
     type: "Range",
     description: "Deleting this range will also delete all child ranges.",
   });
-  const closeTabs = PPanel.useCloseResourceTabs();
   const { update: del } = Ranger.useDelete();
   const { update: renameRange } = Ranger.useRename();
   const handleAddChildRange = () => {
@@ -84,7 +74,6 @@ export const ContextMenu = ({ keys }: Menu.ContextMenuMenuProps) => {
       if (!confirmed) return;
       const keys = ranges.map((r) => r.key);
       dispatch(Session.Range.remove({ keys }));
-      closeTabs(ranger.ontologyID(keys));
       del(keys);
     }, "Failed to delete range");
   };
@@ -92,11 +81,14 @@ export const ContextMenu = ({ keys }: Menu.ContextMenuMenuProps) => {
   return (
     <Base.Menu>
       {isSingle && (
+        <Menu.Item itemKey="details" onClick={handleDetails}>
+          <Icon.Details />
+          View details
+        </Menu.Item>
+      )}
+      <Menu.Divider />
+      {isSingle && (
         <>
-          <Menu.Item itemKey="details" onClick={handleDetails}>
-            <Icon.Details />
-            View details
-          </Menu.Item>
           {hasUpdatePermission && <Base.RenameItem onClick={handleRename} />}
           {hasCreatePermission && (
             <Menu.Item itemKey="addChildRange" onClick={handleAddChildRange}>
@@ -104,35 +96,29 @@ export const ContextMenu = ({ keys }: Menu.ContextMenuMenuProps) => {
               Create child range
             </Menu.Item>
           )}
-          <Divider.Divider x />
         </>
       )}
+      <Menu.Divider />
       <Base.FavoriteItems
         anyFavorited={someAreFavorites}
         anyNotFavorited={someAreNotFavorites}
         onFavorite={handleFavorite}
         onUnfavorite={handleUnfavorite}
       />
-      {(someAreFavorites || someAreNotFavorites) && <Divider.Divider x />}
-      {hasDeletePermission && isNotEmpty && (
-        <>
-          <Base.DeleteItem onClick={handleDelete} />
-          <Divider.Divider x />
-        </>
-      )}
+      <Menu.Divider />
       {isSingle && (
-        <>
-          <Link.CopyContextMenuItem
-            onClick={() =>
-              handleLink({
-                name: ranges[0].name,
-                ontologyID: ranger.ontologyID(ranges[0].key),
-              })
-            }
-          />
-          <Divider.Divider x />
-        </>
+        <Link.CopyContextMenuItem
+          onClick={() =>
+            handleLink({
+              name: ranges[0].name,
+              ontologyID: ranger.ontologyID(ranges[0].key),
+            })
+          }
+        />
       )}
+      <Menu.Divider />
+      {hasDeletePermission && isNotEmpty && <Base.DeleteItem onClick={handleDelete} />}
+      <Menu.Divider />
       <Base.ReloadConsoleItem />
     </Base.Menu>
   );
